Add a release workflow
To make a new release:

1. Click on "Create new release" under the "Releases" section on
https://github.com/TopoToolbox/BankfullMapper.
2. Create a new tag under "Choose a tag" with the new version
number (e.g. v1.0.0) and choose "main" as the target branch.
3. GitHub can generate release notes from the recent PRs, and you can
supplement these with release notes in the text book.
4. Upon clicking "Publish release," the release will be available
under the Releases section on the repository page, and the workflow
will run.
5. In a few minutes, the packaged BankfullMapper.mltbx file will be
available from the release page.

The workflow is copied largely from the TopoToolbox 3 release
workflow. Because BankfullMapper is pure MATLAB, it does not need to
be compiled for each operating system, and we only run it the workflow
Linux. We also trigger the workflow manually rather than
automatically. If a more complex build process is required in the
future (such as needing platform-specific toolboxes), the TopoToolbox
3 and pytopotoolbox release workflows can be used as a reference.

This PR also fixes the name of the test class to match the file so
that the test runs without warnings.